==> Step4_coMet_Tree_NullModel.R <==
#!/usr/bin/env Rscript

suppressPackageStartupMessages(if(!require(cli)){install.packages("cli");require(cli)})
# Display a beginning message. 
cat(rule(left = "SCRIPT COMET_TREE_NULLMODEL.R BEGINNING", line_col = "red", line = "-", col = "br_red")) 

# ---------------------------------------------#
# coMet pipeline: Creation of tree null-models #
# -------------------------------------------- #

# This script takes as input the trees generated by the coMet_ComSim to create a NULL-MODEL by a permutation of the tips. 

# It takes as input a tree file to be randomized and the number of replicates to perform and outputs a multi-phylo object of N replicates trees. 
# It supports launches with REGEXES including multiple tree files, as soon as the number of replicates is correctly entered as the first argument. 

# There is two ways for randomization: 
  # Randomization of the whole meta-community tree. 
  # Randomization of only the tips that are present across all the samples, discarding the tips of species never present. 

# ------------------------------ #
##### STEP 0: INITIALISATION #####
# ------------------------------ #

##### ______________________________________________________________________________________________________________________________________ #####

cat(rule(left = "INITIALISATION", line_col = "green", line = "-", col = "br_green"))

    # ----- #

  #### . Packages . ####

# Print a message
cli_alert_info("Loading the needed packages ... ")

# Install/load pacman. 
suppressPackageStartupMessages(if(!require(pacman)){install.packages("pacman");require(pacman)})
p_load("foreach",
       "phytools",
       "doParallel",
       "dplyr",
       "argparser",
       "tidyr",
       "data.table")

    # ----- #

  #### . Local Mode . ####

Parameters <- list()
Parameters$Config <- "Foo.R"

    # ----- #

  #### . Argument Parser . ####

cli_alert_info("Reading the coMet_ConfigFile.R ...")

# Create the parser
arg_parser <- arg_parser("Check the communities computed by coMet_ComSim and create an output pdf.", name = "coMet_ComSim_Check.R", hide.opts = FALSE)

# Add the config_files.R as positional arguments
arg_parser <- add_argument(arg_parser, arg = "Configuration", nargs = 1, help = "Configuration file for coMet_ComSim.R")

# Parse the arguments
Parameters <- parse_args(parser = arg_parser, argv = commandArgs(trailingOnly = T))

# Print a message to inform that the right configuration file was correctly loaded. 
cli::cli_alert_info(paste0("Working on configuration file: ",Parameters$Config))

# Load them into the environnement
source(Parameters$Config)

# Print a message
cli_alert_success("coMet_ConfigFile.R correctly read !")

    # ----- #

  #### . Loading files . ####

# Display some information about the config-file to be sure that all is OK.
cat(col_yellow("Scenario: "),Scenario,"\n")
cat(col_yellow("Number of Null trees: "),Ntree,"\n")

# Load the correct trees from the multiphylo object 
Trees <- read.tree(paste0(getwd(),"/coMet_ComSim_Outputs/",Scenario,"/Whole_Communities/Scenario",Scenario,"_Step1_Total_PhyloTrees.tree"))

# Get the number of replicates
RepTree <- Ntree


  # The number of trees could not be the wanted number of replicates because "Wrong" 

# Each Scenario will be traited with "CONF"
# The multiphylo object "TREES" is created. 

# Make test
if (round(RepTree) != RepTree) {stop("The first argument is not an integer, please enter an integer as the number of replicates to perform.")}

# ------------------------------------------- #
##### STEP 1: TOTAL PHYLOGENETIC SHUFFLE  #####
# ------------------------------------------- #

if(Randomization == 1){

##### ______________________________________________________________________________________________________________________________________ #####

# Display a message
cat(rule(left = "CREATION OF THE  NULL MODEL PHYLOGENETIC TREES", line_col = "green", line = "~", col = "br_green"))  

    # -----#   

  #### . /!\ Beginning of the loop /!\ . ####

# Create a loop to load all tree replicated for each scenario
NM <- foreach(i = 1:length(Trees)) %dopar% {  # Looping on all of the trees from the multiphylo object tree
  # Load the tree(s)
  Phylo_tree <- Trees[[i]] #; cli_alert_info(paste0("Tree ",i,"/",length(Trees), " correctly loaded"))
 
  #### . Computation of the trees . ####
  
  # Use a foreach loop to compute
  Phylo_NULL <- foreach(j = 1:RepTree) %dopar% {
    # Get a vector of random names from the species names
    Random_Name <- sample(Phylo_tree$tip.label,size = length(Phylo_tree$tip.label))
    # Get the good tree topology
    Phylo_tree_NULL_MODEL <- Phylo_tree
    # Create the RANDOM tree (by putting back the vector of species names shuffled)
    Phylo_tree_NULL_MODEL$tip.label <- Random_Name
    # Return the RANDOM tree. 
    return(Phylo_tree_NULL_MODEL)     
  }


  # Transform the list into a multiphylo_Object
class(Phylo_NULL) <- "multiPhylo"
  # Rename it
names(Phylo_NULL) <- paste0("Tree_NULL_MODEL_",1:RepTree)
  # Save it
write.tree(Phylo_NULL,paste0(getwd(),"/coMet_ComSim_Outputs/",Scenario,"/Whole_Communities/TreeDistances/Scenario",Scenario,"_Step3_MultiPhylo_NullModel",i,".tree"))
  # Return
return(Phylo_NULL)

}

  # -----#   

#### . /!\ End of the loop /!\ . ####

} # END OF RANDOMIZATION == 1

# --------------------------------------------- #
##### STEP 2: SAMPLED PHYLOGENETIC SHUFFLE  #####
# --------------------------------------------- #

if(Randomization == 2){

##### ______________________________________________________________________________________________________________________________________ #####

# Display a message
cat(rule(left = "CREATION OF THE  NULL MODEL PHYLOGENETIC TREES", line_col = "green", line = "~", col = "br_green"))  

  # ----- #

#### . Loading files . ####

# Load the sampled communities
Samples <- read.csv(paste0(getwd(),"/coMet_ComSim_Outputs/",Scenario,"/Sampled_Communities/Scenario",Scenario,"_Step4_Total_Sampled_Occurence.csv"), row.names = 1)

# For each replicates, only keep the species that are present at least one time across all replicates
Sampled_Species <- Samples %>% 
  # Pivot the species
  pivot_longer(
    cols = starts_with("t"), names_to = "Species", values_to = "Occurence", values_drop_na = TRUE) %>%
  # Remove the lines where the specie is absent
  filter(Occurence != 0)

#### . /!\ Beginning of the loop /!\ . ####

# Create a loop to load all tree replicated for each scenario
NM <- foreach(i = 1:length(Trees)) %do% {  # Looping on all of the trees from the multiphylo object tree
  
  # Load the tree(s)
  Phylo_tree <- Trees[[i]]
  
  # Extract the names of the species present at least one time in the replicate
  Species <- Sampled_Species %>%
    dplyr::filter(rep == i) %>% # Choose the replicate
    select(Species) %>% # Select the species
    unique() %>% # Get the species
    unlist()
  
  # Keep the tips in the corresponding tree
  Phylo_tree <- keep.tip(Phylo_tree,Species)
  
  # Save the Trimmed_Observed Tree
  write.tree(Phylo_tree,paste0(getwd(),"/coMet_ComSim_Outputs/",Scenario,"/Whole_Communities/TreeDistances/Scenario",Scenario,"_Step3_ObservedTrimmed_NullModel",i,".tree"))
  
  #### . Computation of the trees . ####
  
  # Use a foreach loop to compute
  Phylo_NULL <- foreach(j = 1:RepTree) %dopar% {
    # Get a vector of random names from the species names
    Random_Name <- sample(Phylo_tree$tip.label,size = length(Phylo_tree$tip.label))
    # Get the good tree topology
    Phylo_tree_NULL_MODEL <- Phylo_tree
    # Create the RANDOM tree (by putting back the vector of species names shuffled)
    Phylo_tree_NULL_MODEL$tip.label <- Random_Name
    # Return the RANDOM tree. 
    return(Phylo_tree_NULL_MODEL)     
  }
  
  # Transform the list into a multiphylo_Object
  class(Phylo_NULL) <- "multiPhylo"
  # Rename it
  names(Phylo_NULL) <- paste0("Tree_NULL_MODEL_",1:RepTree)
  # Save it
  write.tree(Phylo_NULL,paste0(getwd(),"/coMet_ComSim_Outputs/",Scenario,"/Whole_Communities/TreeDistances/Scenario",Scenario,"_Step3_MultiPhylo_NullModel",i,".tree"))
  # Return
  return(Phylo_NULL)
  
}

    # -----#   

  #### . /!\ End of the loop /!\ . ####

} # END OF RANDOMIZATION == 2

# Print a message 
cli_alert_success("Null-Model Phylogenetic trees correctly saved !")  

# END OF THE SCRIPT
cat(rule(left = "SCRIPT COMET_TREE_NULLMODEL.R ENDING", line_col = "red", line = "-", col = "br_red"))
